[jsscripting] Move ModuleLocator & ScriptExtensionModuleProvider into scope sub-package
